Make body length 64-bit; bound getContent() at the array ceiling
Body length was reported by an int getLength(): Body/ChunkedBody/Http2Body
cast (int)streamedLength and AbstractBody summed chunk lengths as int. For a
body >= 2 GB this silently overflowed to a negative value (exactly 2^31 bytes
-> Integer.MIN_VALUE). Since #3039 tightened Header.getContentLength() to
reject negative/non-numeric values, that overflowed length written via
setContentLength() then threw MalformedHeaderException during response
construction (e.g. ReturnInterceptor.createResponseFromRequest).

The wire/streaming path is already 64-bit clean (Header content-length and
Body(in, long length) use long, writeStreamed streams >2 GB fine); only the
reported length API and getContent() materialization broke.

- getLength() returns long across AbstractBody, Body, ChunkedBody, EmptyBody,
  the SSE StreamingBody and HTTP/2 Http2Body (streamedLength field widened),
  removing the (int) truncations.
- getContent() throws the new BodyTooLargeException before allocating when the
  length exceeds MAX_ARRAY_LENGTH (Integer.MAX_VALUE - 8) instead of
  overflowing/OOMing; same guard on ChunkedBody.getRawLength(). A single byte[]
  cannot hold more than ~2 GB, so full-body materialization stays small-message
  only. getContentAsStream() is unaffected (wraps the chunk list).
- BodyTooLargeException extends ReadingBodyException so existing body-error
  handling keeps working while the subtype stays distinguishable.
- Narrowing consumers fixed: Message.estimateHeapSize() clamps; two
  out.write(getContent(), 0, getLength()) sites use the array's own length.

+++ core/src/main/java/com/predic8/membrane/core/interceptor/sse/ServerSentEventsDemoStreamInterceptor.java
@@ -108,7 +108,7 @@ public class ServerSentEventsDemoStreamInterceptor extends AbstractInterceptor {
         }
 
         @Override
-        public int getLength() {
+        public long getLength() {
             return -1; // Unknown length due to streaming
         }
